+++ CHANGELOG.md
@@ -9,6 +9,9 @@ All notable changes to this project will be documented in this file. This change
  - Ability to update bot's status on Discord
  - Ability to update bot's voice state on Discord
  - Ability to get roles from a guild using the messaging api
+ 
+### Fixed
+ - Improper handling of rate limits when none is found
 
 ### Removed
  - Uberjar profile in Leiningen